Refactor: Decompose ProductivityScreen God Class
Extracted the massive UI block from `frontend/app/(main)/productivity.tsx` into four distinct components to adhere to SOLID principles and the 50-line sub-widget rule constraint without altering behavior:
- ProductivityHeader
- ProductivityTabs
- ProductivityEmptyState
- ProductivityTodayPlan
